change dir structure; add MyToken example as erc20 impl
