+++ src/main.rs
@@ -1,3 +1,5 @@
+mod weth_example;
+
 pub fn main() {
     
 }
