+++ src/tokens/mod.rs
@@ -1,0 +1,2 @@
+pub mod erc20;
+pub mod my_token;
